refactor: Rename SYSTEM_PROMPT_TEMPLATE to START_MESSAGE_TEMPLATE and change role to user
The `role` of the first message in a conversation must be `user`, this
commit changes the role from `system` to `user` to align with this
requirement.

Additionally, this commit introduces a `fix_input_schema` function that
removes the `format: uri` from the input schema of tools. This is a
workaround for the fact that Gemini API does not support this field.